Fix BNF Stringify Visitor recursion case
If we've already seen a parser but it doesn't have a name, we can still visit it if it has no children. Having 0 children before we visit it again helps prevent loops. If it has children, we'll print the ALREADY SEEN UNNAMED PARSER error message like usual
Add unit tests for both cases