Initial implementation of a Regex parser
Regex(pattern) takes a pattern, and tries to match that pattern using normal Regular Expression rules from the current position in the input stream
RegexParser implements the basics of a regex parsing engine. Some cleanups and optimizations TBD.
RegexParserMethods implements the grammar for regex patterns and a wrapper method to turn a pattern into a RegexParser
Current implementation is two steps: One to parse the pattern into a tree of IRegexNodes and a second to linearize that tree into a list of RegexState. The engine operates on the list of Regex state.
Current implementation supports several important regex features. Additional features may wait until a later release.