Work on recursive state data, with quick XML parser example
Add a new CascadingKeyValueStore which is a linked list of dictionaries. Data values can be accessed in the current or previous frames by name.
Add IParser extension methods to work with state: a method to add the result of an inner parser to state. More are probably needed
Fix a Bug in RuleParser where we were creating an extra ParseState<T>
Add a new CreateParser and Create() parser method to create a parser at parse time using information from the current parse state (including input and data).
CreateParser is very similar to ChainParser, and we might be able to unify the two in the future
Add an XML parser to the test examples, which shows the use of stateful data to parse matching xml tags. More tests are needed but it's a good start