Rewrite the Pratt parser to use a better, more flexible and powerful implementation of the algorithm.
Instead of hard-coding prefix/postfix/whateverfix operators in lists, delegate to a list of configured parselets which hold the necessary information and callbacks that handle each specific case.
Every parselet returns an internal Token type, which holds all rules in an operator-agnostic way.
A context object is passed to the parser, which allows user-callbacks to recurse into the parser engine to gather more values.
Several unit tests and examples had to be modified to allow for a more expressive fluent builder
Add a unit test for operators with mixed association, to show that it works as expected
Keeping everything wrapped inside the Pratt<TInput, TOutput> parent class, because TOutput and TInput are omnipresent throughout all this code, and having to specify those type parameters on every single type is very cumbersome. Plus it gives us a good hiding mechanism for some of the internal details that otherwise would be publicly visible
At least break the Pratt class into separate files, and move some of them to a separate folder