Parsers which loop or recurse will now detect and handle zero-length inputs
Pratt will not continue looping on right-hand rules if the parselet has zero-length
Add a TODO note about stopping recursion on left-hand rules if a parselet returns zero-length. We'll have to address that (and add tests for both cases) later
In some cases a zero-length value can be tolerated, and in some places it causes failure.
Add unit tests for these cases.