Add a Try parser
TryParser catches user exceptions. On exception, it rewinds the input sequence to the starting point of Try, and then gives the user the option whether to return a failure result or to allow the exception to bubble upwards. The TryParser does not interfere with ControlFlowExceptions, with some unit tests showing this general operation.
Add an overload for Fail() which does not specify an explicit output type (the input type is used as the output type, like Match())