Add a Cache parser to be able to cache results by (parser, location).
Add a Null cache (default) and an IMemoryCache-based implementation.
More are possible but not currently provided.
ICacheStatistics can be used to get basic usage information from the cache.
Cache is exposed in ParseState.Cache
Cache(p) can be used to cache the output of parser p. Add a few parser methods and extension methods to support.