Fix StyleCop rule
I can't enable IDE0090 because VisualStudio is autoformatting 'new(...)' instead of 'new (...)' because it's treating it like a method invocation which should not have a space between the method name and the open paren. This even though SA1000 requires having a space after a keyword. If there's a way to fix this spacing issue only for these cases of the new keyword only, I haven't found it yet