Re-implement StreamCharacterSequence to use stream rewind instead of caching all stream contents
We can use stream.Seek() and _reader.DiscardBufferedData() to randomly seek inside the stream, so long as we keep track of stream position at each point. StreamReader doesn't keep track of this for us (in a way that is exposed) so we have to do it ourself with Encoding.GetByteCount(char*) which requires unsafe code for performance.
For each character read, we increment the count of chars consumed and we also increase the current stream position by the size of that character according to that encoding
If we see a High Surrogate we MUST see a Low Surrogate immediately to calculate byte size and advance the stream position. reported byte sizes are invalid for a single surrogate.
Checkpoints are disabled between surrogate pairs in this sequence
When we .GetNext() after the high surrogate we return the low surrogate, clear the flag, but do not advance the stream position again.
Use a struct for metadata so we can use a single assignment statement to copy over all values. This helps to keep the rewind mechanism cleaner and easier to modify
All existing tests pass, but we need to start adding some tests for multi-byte characters
Some TODO notes added. Some cleanup required. More test coverage would be nice.