Initial implementation of a panic-mode 'Synchronize' parser
This parser attempts the parse and returns immediately if the result is a success
If the initial attempt does not succeed, the parser enters 'panic mode' where it attempts to discard input tokens and continue
This is handy in cases where you want to display several parse-errors to the end user, not just fail and stop at the first bad input.
After the initial attempt fails, the parser will always return failure, even if it eventually succeeds.
All error results are recorded in result.TryGetData<ErrorList>()
If the parser eventually succeeds, the final success will be recorded in result.TryGetData<IResult>()