Add Optional parser classes (again)
Optional() had been working by using TransformResult() to turn error results into success results. But this caused a problem with the optimized Match() implementation of Transform parser.
For that reason, I had to add Optional parser classes, which can work with .Match() in a correct and meaningful way.
Maybe we need to add a Match callback to Transform()?
Add several notes throughout to explain to future-me why we can't optimize what we want to.
We might also be able to add a Transform variant for cases where we are just modifying a result value (which we could optimize)