Several test fixes
Fix Try() tests that we depending on old Produce().Match() behavior. Produce().Match() is optimized to always return true now, even if the user callback would have thrown an exception.
Fix an error where the BnfStringify visitor Accept() mechanisms can't dispatch when one of the type params is private and not visible from the visitor. Make that public.
Move Function().ToBnf() tests to the same file as the other Function() tests