s/PrenormalizedStringCharacterSequence/CharArraySequence/
This type is kind of general-purpose. We convert strings or IReadOnlyList<char> to char[] and work on it that way
Add option to not normalize. The Nonnormalized version optimizes by not normalizing and storing the string in-place, but we can still avoid normalization with this type
Add a first set of tests for it