Sequences which read values from a user-provided Delegate method.
DelegateSource and DelegateCharSource read values from a specially-crafted user delegate type.
The delegate method returns a ValueTuple (T next, bool atEnd), where the 'next' is the item read and `atEnd` returns true if the sequence is at or passed the end *after the read*.
These two sources use double-buffering to try and minimize the number of calls to the delegate.
Added a new .ItemsGenerated to SequenceStatistics to track the number of times we call the delegate.
There is a pathological case where rewinding to the end of an unloaded buffer may cause the entire buffer to be refilled in order to only read one or two values. A LowWaterMark for the buffer may help with this, if it comes up often, but raises it's own concerns.
Docs for these, and for sequence statistics.