De-duplicate code in Chain parser
We can use a struct with parser/result type information to hold most of the logic of Chain. Then we can delegate to it.
Also merge and simplify classes a little bit.
There are a few issues where we now have to prepend Internal.Parsers.Chain<> to not conflict with Chain() factory method. Renaming the class would help, but I'm blanking on ideas right now