Re-implement BNF stringification using a new visitor design
Structured visitors now consist of four parts (some of which may be the same object, depending on implementation):
1. A .Visit() method on IParser with a double-dispatch which allows each parser to select it's own strongly-typed Accept variants
2. The visitor State, which is passed along during the visit to each parser and either is the visitor itself or contains reference to it
3. The IPartialVisitor<TState> which implements the individual .Accept() method variants for each supported parser typed
4. The IVisitor<TState> which has a method to get the appropriate IPartialVisitor<TState> for each parser on request.

The parser .Visit() method calls "visitor.Get<PartialVisitorType>()?.Accept(this, state)"
If a matching partial visitor doesn't exist, the parser simply does nothing.
We have pre-defined IPartialVisitor interfaces for all internal types, and downstream libraries can provide their own interfaces for their own custom parser types
The partial visitor in turn can call some kind of method on TState (for BNF stringification, it's the .Append() method)
The TState can then call "parser.Visit(_visitor, this)" to move to the next parser.

I would like to consider a design which uses trampolines instead of endless recursion, but for now this is the design we're going with.
Remove IHiddenInternalParser interface, it is no longer required.