Remove several allocations from Regex
RegexMatch IS-A dictionary instead of HAS-A dictionary. this saves an allocation
Allocate the dictionary as <int, IReadOnlyList<string>> so we don't have to allocate two dictionaries (one with List<string> and the other with IReadOnlyList<string>)
Allocate arrays instead of lists, where possible
Allocate and build the dictionary directly instead of allocating an IReadOnlyList<(int, string)> first.
Overall this should be at least 3 fewer allocations per successful Regex parse