Start re-working TryGetData
ResultData now takes an object. If object is T return it directly. If object is IReadOnlyList<object> iterate and find a matching T
We're optimizing for the case that creating these results is common, but actually reading them is relatively rare
A few IParseState<T> methods now take a ResultData object directly instead of a list of object. This saves us on allocating several arrays throughout for one item
Remove data options from a few method signatures where we've never even tested it