Allow same-buffer rewinds for StreamCharacterSequence
Realize that the first character in the buffer has .Consumed value N which is always a multiple of BufferSize M (N = 0M, 1M, 2M, ...)
So at any point we can find the Consumed value for the beginning of the buffer with simple integer arithmetic
Assuming that by default buffers are pretty big, we can allocate an array with a number of items to hold BufferSTartPosition offsets for each seen buffer
On Rewind() we can find where the start position of the stream should be for that buffer, rewind to it.
In the case where CurrentBufferStart == DesiredBufferStart we can avoid a refill entirely.
As usual, caveat about a case where we frequently read/rewind around a buffer boundary leading to pathological refills. With this optimization in place it might make sense to consider double-buffering again.