MatchChar(predicate) parser
Match() with a predicate can match an end sentinel. However in some repetition situations we want to not include the end sentinel in the list (Repetition will break after a single read where Consumed==0, so we get one sentinel). For char situations we don't want \0 so add a MatchChar(predicate) that excludes end sentinels and fails if the input is at end.
Add a flag to MatchPredicateParser to switch between the two behaviors
Add a few tests, both to codify the existing behavior of Match(predicate) and demonstrate the new behaviors of MatchChar(predicate)
We probably want a new parser which uses the new behavior but isn't char-specific, but I haven't figured out what to call it yet.