Rule Parser can rent an array when flagged
We know in many situations the results array is discarded immediately after the result value is created. This is true in all Rule() method variants and false in all Combine() method variants, etc.
So we add a flag to Rule.Parser which tells us whether we need to allocate the array on the heap or we can rent it from the ArrayPool
Update all usages.