Reduce allocations in Trie implementation
Instead of a Node containing a .Value, the parent Node keeps a tuple of (childNode, hasValue, value). This means we can store the value in the existing parent node and not have to allocate leaf nodes.
Some of the logic gets a bit more complicated and there are some opportunities to streamline, but not having to allocate leaf nodes is a net win