wip Start implementing some parser base classes.
I know, I know, I hate inheritance. I'm doing the experiment anyway to see if the reduction in lines of code is significant enough for me to suck it up and deal. So far, some of the simple parser types are showing some promising shortening, but I will have to see if I want to continue with the remainder.