Rework Compose() parser
Rule.Parser can now specify an item type and is abstracted to be able to work on outputless and outputted parsers.
If all parsers have the same output type we can avoid the cast/box to object.
This is a small optimization because Compose() is used less often than Rule() and Rule() is source-generated to be a little more optimal with tuples instead of arrays/lists for intermediate results.
But this is still worthwhile because we have a few places that use Compose() and we want those to be speedy.
Also this parser doesn't play with the ArrayPool. For small arrays like we are (probably) using here, it isn't a savings to use the arraypool.