Merge Examine into Context
Add in the .Consumed tracking to Context. Remove comments about how we expect Context not to be adjusting the state. If the user has access to mutable state, we must expect them to adjust it.
Add a Context() variant for IParser<TInput>
Context() callback methods use the ParseContext<TInput, TOutput> struct instead of just the raw parse state. This gives us a place to add additional functionality in the future.