Cleanup Select, Chain, ChainWith
Select does not need a struct with an extra level of indirection. ResultAlternative<> is a struct now, so we can just return default and see that .Success==false
Also add a delegate type for select
Remove Chain() and ChainWith() parsers which take IParser<TInput> as the first argument. How can we select a parser based on the result value, if we don't have any idea what the result value may or may not be? At least we should take IParser<TInput, object>.