Remove DataFrame parser and implement with Context()
Add IParseState<> methods to .PushDataFrame() and .PopDataFrame() so we can do them separately.
Also add a .GetCurrentDataFrame() which Try() uses to re-set the data frame to the point it was at when the Try() was entered
Otherwise if we throw, we may want to examine the state of the data frame at the point when the exception was thrown, so keep it around. Once we fail the parse, it doesn't matter if the data frame is cleaned.