Initial implementation of Regex recursing into other parsers
Right now we store a list of named parsers into IParseState<char>.Data. Then when we see a (?pname) named parser in the regex pattern we search for an IParser<char> in the context.Data and use that
We can recurse into the parser.Match method because RegexContext ISA IParseState<char>
I am not entirely happy with the implementation because we're churning up a lot of garbage to create this. Need to refactor with an eye towards simplifying and optimizing